Moved simulation MHE into its own file
