+++ examples/moving_horizon_estimation/mhe.py
@@ -1,15 +1,10 @@
-import importlib
-from pathlib import Path
-
-import biorbd
-import numpy as np
 import time
 
-# Load pendulum
-FILE_FOLDER = Path(__file__).parent
-spec = importlib.util.spec_from_file_location("mhe_simulation", str(FILE_FOLDER) + "/mhe_simulation.py")
-mhe_simulation = importlib.util.module_from_spec(spec)
-spec.loader.exec_module(mhe_simulation)
+import matplotlib.pyplot as plt
+from scipy.integrate import solve_ivp
+import casadi as cas
+import numpy as np
+import biorbd
 
 from biorbd_optim import (
     Instant,
@@ -27,6 +22,90 @@ from biorbd_optim import (
     Data,
     Solver,
 )
+
+
+def generate_data(biorbd_model, Tf, X0, T_max, N, noise_std, SHOW_PLOTS=False):
+    # Casadi functions
+    x = cas.MX.sym("x", biorbd_model.nbQ() + biorbd_model.nbQdot())
+    u = cas.MX.sym("u", 1)
+    q = cas.MX.sym("q", biorbd_model.nbQ())
+
+    forw_dyn = cas.Function(
+        "forw_dyn",
+        [x, u],
+        [
+            cas.vertcat(
+                x[biorbd_model.nbQ() :],
+                biorbd_model.ForwardDynamics(
+                    x[: biorbd_model.nbQ()], x[biorbd_model.nbQ() :], cas.vertcat(u, 0)
+                ).to_mx(),
+            )
+        ],
+    ).expand()
+    pendulum_ode = lambda t, x, u: forw_dyn(x, u).toarray().squeeze()
+
+    markers_kyn = cas.Function("makers_kyn", [q], [biorbd_model.markers(q)]).expand()
+
+    # Simulated data
+    h = Tf / N
+    # U_ = (np.random.rand(N)-0.5)*2*T_max # Control trajectory
+    U_ = (-np.ones(N) + np.sin(np.linspace(0, Tf, num=N))) * T_max  # Control trajectory
+    X_ = np.zeros((biorbd_model.nbQ() + biorbd_model.nbQdot(), N))  # State trajectory
+    Y_ = np.zeros((3, biorbd_model.nbMarkers(), N))  # Measurements trajectory
+
+    for n in range(N):
+        sol = solve_ivp(pendulum_ode, [0, h], X0, args=(U_[n],))
+        X_[:, n] = X0
+        Y_[:, :, n] = markers_kyn(X0[: biorbd_model.nbQ()])
+        X0 = sol["y"][:, -1]
+    X_[:, -1] = X0
+    Y_[:, :, -1] = markers_kyn(X0[: biorbd_model.nbQ()])
+
+    # Simulated noise
+    np.random.seed(42)
+    N_ = (np.random.randn(3, biorbd_model.nbMarkers(), N) - 0.5) * noise_std
+    Y_N_ = Y_ + N_
+
+    if SHOW_PLOTS:
+        q_plot = plt.plot(X_[: biorbd_model.nbQ(), :].T)
+        dq_plot = plt.plot(X_[biorbd_model.nbQ() :, :].T, "--")
+        plt.legend(
+            q_plot + dq_plot,
+            [i.to_string() for i in biorbd_model.nameDof()] + ["d" + i.to_string() for i in biorbd_model.nameDof()],
+        )
+        plt.title("Real position and velocity trajectories")
+        plt.figure()
+        marker_plot = plt.plot(Y_[1, :, :].T, Y_[2, :, :].T)
+        plt.legend(marker_plot, [i.to_string() for i in biorbd_model.markerNames()])
+        plt.gca().set_prop_cycle(None)
+        marker_plot = plt.plot(Y_N_[1, :, :].T, Y_N_[2, :, :].T, "x")
+        plt.title("2D plot of markers trajectories + noise")
+        plt.show()
+
+    return X_, Y_, Y_N_, np.vstack([U_, np.zeros((N,))])
+
+
+def check_results(biorbd_model, N, Xs):
+    # Casadi functions
+    x = cas.MX.sym("x", biorbd_model.nbQ() + biorbd_model.nbQdot())
+    u = cas.MX.sym("u", 1)
+    q = cas.MX.sym("q", biorbd_model.nbQ())
+
+    markers_kyn = cas.Function("makers_kyn", [q], [biorbd_model.markers(q)]).expand()
+    Y_est = np.zeros((3, biorbd_model.nbMarkers(), N))  # Measurements trajectory
+
+    for n in range(N):
+        Y_est[:, :, n] = markers_kyn(Xs[: biorbd_model.nbQ(), n])
+
+    return Y_est
+
+
+def plot_true_X(q_to_plot):
+    return X_[q_to_plot, :]
+
+
+def plot_true_U(q_to_plot):
+    return U_[q_to_plot, :]
 
 
 def warm_start_mhe(data_sol_prev):
@@ -104,14 +183,6 @@ def prepare_ocp(
     )
 
 
-def plot_true_X(q_to_plot):
-    return X_[q_to_plot, :]
-
-
-def plot_true_U(q_to_plot):
-    return U_[q_to_plot, :]
-
-
 if __name__ == "__main__":
 
     biorbd_model_path = "./cart_pendulum.bioMod"
@@ -125,7 +196,7 @@ if __name__ == "__main__":
     N_mhe = 10  # size of MHE window
     Tf_mhe = Tf / N * N_mhe  # duration of MHE window
 
-    X_, Y_, Y_N_, U_ = mhe_simulation.run_simulation(biorbd_model, Tf, X0, T_max, N, noise_std, SHOW_PLOTS=False)
+    X_, Y_, Y_N_, U_ = generate_data(biorbd_model, Tf, X0, T_max, N, noise_std, SHOW_PLOTS=False)
 
     X0 = np.zeros((biorbd_model.nbQ() * 2, N_mhe + 1))
     X0[:, 0] = np.array([0, np.pi / 2, 0, 0])
@@ -180,9 +251,8 @@ if __name__ == "__main__":
     print(f"Average time per iteration of MHE : {(t1-t0)/(N-N_mhe-2)} s.")
     print(f"Norm of the error on state = {np.linalg.norm(X_[:,:-N_mhe] - X_est)}")
 
-    Y_est = mhe_simulation.check_results(biorbd_model, N - N_mhe, X_est)
+    Y_est = check_results(biorbd_model, N - N_mhe, X_est)
     # Print estimation vs truth
-    import matplotlib.pyplot as plt
 
     plt.plot(Y_N_[1, :, : N - N_mhe].T, Y_N_[2, :, : N - N_mhe].T, "x", label="markers traj noise")
     plt.gca().set_prop_cycle(None)
